+++ src/java/jsonista/jackson/PersistentHashMapDeserializer.java
@@ -6,32 +6,48 @@ import clojure.lang.PersistentHashMap;
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.JsonToken;
-import com.fasterxml.jackson.databind.DeserializationContext;
-import com.fasterxml.jackson.databind.JavaType;
-import com.fasterxml.jackson.databind.JsonDeserializer;
-import com.fasterxml.jackson.databind.KeyDeserializer;
+import com.fasterxml.jackson.databind.*;
+import com.fasterxml.jackson.databind.deser.ContextualDeserializer;
 import com.fasterxml.jackson.databind.deser.std.StdDeserializer;
 
 import java.io.IOException;
 import java.util.Map;
 
-public class PersistentHashMapDeserializer extends StdDeserializer<Map<String, Object>> {
+public class PersistentHashMapDeserializer extends StdDeserializer<Map<String, Object>> implements ContextualDeserializer {
+
+  private KeyDeserializer _keyDeserializer;
+  private JsonDeserializer<?> _valueDeserializer;
 
   public PersistentHashMapDeserializer() {
     super(Map.class);
+  }
+
+  public PersistentHashMapDeserializer(KeyDeserializer keyDeser, JsonDeserializer<?> valueDeser) {
+    this();
+    _keyDeserializer = keyDeser;
+    _valueDeserializer = valueDeser;
+  }
+
+  protected PersistentHashMapDeserializer withResolved(KeyDeserializer keyDeser, JsonDeserializer<?> valueDeser) {
+    return this._keyDeserializer == keyDeser && this._valueDeserializer == valueDeser ? this : new PersistentHashMapDeserializer(keyDeser, valueDeser);
+  }
+
+  @Override
+  public JsonDeserializer<Map<String, Object>> createContextual(DeserializationContext ctxt, BeanProperty beanProperty) throws JsonMappingException {
+    JavaType object = ctxt.constructType(Object.class);
+    KeyDeserializer keyDeser = ctxt.findKeyDeserializer(object, null);
+    JsonDeserializer<Object> valueDeser = ctxt.findNonContextualValueDeserializer(object);
+    return this.withResolved(keyDeser, valueDeser);
   }
 
   @Override
   @SuppressWarnings("unchecked")
   public Map<String, Object> deserialize(JsonParser p, DeserializationContext ctxt) throws IOException, JsonProcessingException {
     ITransientMap t = PersistentHashMap.EMPTY.asTransient();
-    JavaType object = ctxt.constructType(Object.class);
-    KeyDeserializer keyDeser = ctxt.findKeyDeserializer(object, null);
-    JsonDeserializer<Object> valueDeser = ctxt.findNonContextualValueDeserializer(object);
     while (p.nextToken() != JsonToken.END_OBJECT) {
-      Object key = keyDeser.deserializeKey(p.getCurrentName(), ctxt);
+      Object key = _keyDeserializer.deserializeKey(p.getCurrentName(), ctxt);
       p.nextToken();
-      Object value = valueDeser.deserialize(p, ctxt);
+      Object value = _valueDeserializer.deserialize(p, ctxt);
       t = t.assoc(key, value);
     }